Defend against numerical instability when computing number of clusters
It appears some transcendental functions like `log` will yield imprecise which
after the call to `ceil` end up increasing exact logarithms like log(6; 216) by
one and thereby the number of clusters ends up one as well. This means the
algorithm enters infinite recursion and eventually overflows the stack.

This change defends against this by always creating at least two clusters if the
compution yields one. Furthermore, since the number of clusters is neither
guaranteed nor required to be exact (as long as it is not pathological), it also
relaxed the precision from double to single for the efficiency gain.